Add uint8 support for mixed sparse tensor patterns
- Add --uint8 flag to generate_synthetic_patterns.py for memory efficiency
- Implement uint8 patterns tensor generation with [0,255] value scaling
- Confirm x-coordinate alignment to multiple of 4 for CUDA vectorization
- Fix visualization tool CuPy import and array conversion issues
- Update SingleBlockMixedSparseTensor to_array() method for proper dtype handling
- Add dtype-aware normalization in visualization tool for both uint8 and float32
- Replace non-existent extract_pattern() with to_array() method
- Add robust CuPy/NumPy array conversion helper

+++ src/utils/single_block_sparse_tensor.py
@@ -418,7 +418,8 @@ class SingleBlockMixedSparseTensor:
             Dense array, shape (height, width)
         """
         # Create dense array and place block (all blocks assumed to be set)
-        dense = cp.zeros((self.height, self.width), dtype=cp.float32)
+        # Use the same dtype as the stored data
+        dense = cp.zeros((self.height, self.width), dtype=self.dtype)
 
         top_row = int(self.block_positions[channel_idx, batch_idx, 0])  # Use channels-first indexing
         top_col = int(self.block_positions[channel_idx, batch_idx, 1])  # Use channels-first indexing
